moving helper functions to utils && remove unused functions
