+++ rai/results/results.go
@@ -24,311 +24,7 @@ import (
 	"github.com/apache/arrow/go/v9/arrow"
 	"github.com/apache/arrow/go/v9/arrow/array"
 	"github.com/relationalai/rai-sdk-go/rai"
-	"github.com/relationalai/rai-sdk-go/rai/pb"
 )
-
-func mapValueType(typeDef map[string]interface{}) (map[string]interface{}, error) {
-	slice := 3
-	if len(typeDef["typeDefs"].([]interface{})) < 3 {
-		slice = len(typeDef["typeDefs"].([]interface{}))
-	}
-	var relNames []map[string]interface{}
-	for _, typeDef := range typeDef["typeDefs"].([]interface{})[0:slice] {
-		if typeDef.(map[string]interface{})["type"] == "Constant" &&
-			typeDef.(map[string]interface{})["value"].(map[string]interface{})["type"] == "String" {
-			relNames = append(relNames, typeDef.(map[string]interface{}))
-		}
-	}
-
-	if len(relNames) != 3 ||
-		!(relNames[0]["value"].(map[string]interface{})["value"] == "rel" &&
-			relNames[1]["value"].(map[string]interface{})["value"] == "base") {
-		return typeDef, nil
-	}
-
-	standardValueType := relNames[2]["value"].(map[string]interface{})["value"].(string)
-	switch standardValueType {
-	case "DateTime":
-		return _unmarshall(fmt.Sprintf(`{"type":"%v"}`, standardValueType))
-	case "Date":
-		return _unmarshall(fmt.Sprintf(`{"type":"%v"}`, standardValueType))
-	case "Year":
-		return _unmarshall(fmt.Sprintf(`{"type":"%v"}`, standardValueType))
-	case "Month":
-		return _unmarshall(fmt.Sprintf(`{"type":"%v"}`, standardValueType))
-	case "Week":
-		return _unmarshall(fmt.Sprintf(`{"type":"%v"}`, standardValueType))
-	case "Day":
-		return _unmarshall(fmt.Sprintf(`{"type":"%v"}`, standardValueType))
-	case "Hour":
-		return _unmarshall(fmt.Sprintf(`{"type":"%v"}`, standardValueType))
-	case "Minute":
-		return _unmarshall(fmt.Sprintf(`{"type":"%v"}`, standardValueType))
-	case "Second":
-		return _unmarshall(fmt.Sprintf(`{"type":"%v"}`, standardValueType))
-	case "Millisecond":
-		return _unmarshall(fmt.Sprintf(`{"type":"%v"}`, standardValueType))
-	case "Microsecond":
-		return _unmarshall(fmt.Sprintf(`{"type":"%v"}`, standardValueType))
-	case "Nanosecond":
-		return _unmarshall(fmt.Sprintf(`{"type":"%v"}`, standardValueType))
-	case "FilePos":
-		return _unmarshall(fmt.Sprintf(`{"type":"%v"}`, standardValueType))
-	case "Missing":
-		return _unmarshall(fmt.Sprintf(`{"type":"%v"}`, standardValueType))
-	case "Hash":
-		return _unmarshall(fmt.Sprintf(`{"type":"%v"}`, standardValueType))
-	case "FixedDecimal":
-		typeDefs := typeDef["typeDefs"].([]interface{})
-		td3 := typeDefs[3].(map[string]interface{})
-		td4 := typeDefs[4].(map[string]interface{})
-
-		if len(typeDefs) == 6 &&
-			td3["type"].(string) == "Constant" &&
-			td4["type"].(string) == "Constant" {
-			bits := td3["value"].(map[string]interface{})["value"].(int64)
-			places := td4["value"].(map[string]interface{})["value"].(int64)
-
-			if bits == 16 || bits == 32 || bits == 64 || bits == 128 {
-				return _unmarshall(fmt.Sprintf(`{"type":"Decimal%v", "places":"%v"}`, bits, places))
-			}
-
-			break
-		}
-	case "Rational":
-		{
-			typeDefs := typeDef["typeDefs"].([]interface{})
-			if len(typeDefs) == 5 {
-				tp := typeDefs[3].(map[string]interface{})
-
-				switch tp["type"] {
-				case "Int8":
-					return _unmarshall(`{"type":"Rational8"}`)
-				case "Int16":
-					return _unmarshall(`{"type":"Rational16"}`)
-				case "Int32":
-					return _unmarshall(`{"type":"Rational32"}`)
-				case "Int64":
-					return _unmarshall(`{"type":"Rational64"}`)
-				case "Int128":
-					return _unmarshall(`{"type":"Rational128"}`)
-				}
-			}
-		}
-	}
-
-	return typeDef, nil
-}
-
-func walkTypeDefs(typeDef map[string]interface{}, values []interface{}) (interface{}, []interface{}) {
-	switch typeDef["type"] {
-	case "ValueType":
-		v := values
-		var r interface{}
-		var res []interface{}
-		for _, tp := range typeDef["typeDefs"].([]interface{}) {
-			r, v = walkTypeDefs(tp.(map[string]interface{}), v)
-			res = append(res, r)
-		}
-		return res, nil
-	case "Rational8", "Rational16", "Rational32", "Rational64", "Rational128":
-		return values[0:2], values[2:]
-	default:
-		if typeDef["type"] != "Constant" {
-			return values[0:1][0], values[1:]
-		}
-	}
-	return nil, nil
-}
-
-func unflattenConstantValue(typeDef map[string]interface{}, value []*pb.PrimitiveValue) []interface{} {
-	var values []interface{}
-	for _, arg := range value {
-		values = append(values, mapPrimitiveValue(arg))
-	}
-
-	res, v := walkTypeDefs(typeDef, values)
-	if v != nil {
-		panic("Left values from walkTypeDefs: something went wrong !")
-	}
-
-	return res.([]interface{})
-}
-
-func mapPrimitiveValue(val *pb.PrimitiveValue) interface{} {
-	switch val.Value.(type) {
-	case *pb.PrimitiveValue_StringVal:
-		return string(val.Value.(*pb.PrimitiveValue_StringVal).StringVal)
-	case *pb.PrimitiveValue_CharVal:
-		return val.Value.(*pb.PrimitiveValue_CharVal).CharVal
-	case *pb.PrimitiveValue_BoolVal:
-		return val.Value.(*pb.PrimitiveValue_BoolVal).BoolVal
-	case *pb.PrimitiveValue_Int8Val:
-		return val.Value.(*pb.PrimitiveValue_Int8Val).Int8Val
-	case *pb.PrimitiveValue_Int16Val:
-		return val.Value.(*pb.PrimitiveValue_Int16Val).Int16Val
-	case *pb.PrimitiveValue_Int32Val:
-		return val.Value.(*pb.PrimitiveValue_Int32Val).Int32Val
-	case *pb.PrimitiveValue_Int64Val:
-		return val.Value.(*pb.PrimitiveValue_Int64Val).Int64Val
-	case *pb.PrimitiveValue_Int128Val:
-		return []uint64{val.Value.(*pb.PrimitiveValue_Int128Val).Int128Val.Lowbits, val.Value.(*pb.PrimitiveValue_Int128Val).Int128Val.Highbits}
-	case *pb.PrimitiveValue_Uint8Val:
-		return val.Value.(*pb.PrimitiveValue_Uint8Val).Uint8Val
-	case *pb.PrimitiveValue_Uint16Val:
-		return val.Value.(*pb.PrimitiveValue_Uint16Val).Uint16Val
-	case *pb.PrimitiveValue_Uint32Val:
-		return val.Value.(*pb.PrimitiveValue_Uint32Val).Uint32Val
-	case *pb.PrimitiveValue_Uint64Val:
-		return val.Value.(*pb.PrimitiveValue_Uint64Val).Uint64Val
-	case *pb.PrimitiveValue_Uint128Val:
-		return []uint64{val.Value.(*pb.PrimitiveValue_Uint128Val).Uint128Val.Lowbits, val.Value.(*pb.PrimitiveValue_Uint128Val).Uint128Val.Highbits}
-	case *pb.PrimitiveValue_Float16Val:
-		return val.Value.(*pb.PrimitiveValue_Float16Val).Float16Val
-	case *pb.PrimitiveValue_Float32Val:
-		return val.Value.(*pb.PrimitiveValue_Float32Val).Float32Val
-	case *pb.PrimitiveValue_Float64Val:
-		return val.Value.(*pb.PrimitiveValue_Float64Val).Float64Val
-
-	default:
-		panic(fmt.Sprintf("unhandled metadata primitive value %T", val.Value))
-	}
-}
-
-func getColDefFromProtobuf(reltype *pb.RelType) (map[string]interface{}, error) {
-	if reltype.Tag == pb.Kind_CONSTANT_TYPE &&
-		reltype.ConstantType.Value != nil &&
-		reltype.ConstantType.RelType != nil {
-
-		typeDef, err := getColDefFromProtobuf(reltype.ConstantType.RelType)
-
-		if err != nil {
-			return nil, err
-		}
-
-		if typeDef["type"] != "ValueType" {
-			var values []interface{}
-			for _, arg := range reltype.ConstantType.Value.Arguments {
-				values = append(values, mapPrimitiveValue(arg))
-			}
-
-			var value interface{}
-			var err error
-			if len(values) == 1 {
-				value, err = convertValue(typeDef, values[0])
-			} else {
-				value, err = convertValue(typeDef, values)
-			}
-
-			if err != nil {
-				return nil, err
-			}
-
-			// add value to typeDef
-			typeDef["value"] = value
-			return map[string]interface{}{
-				"type":  "Constant",
-				"value": typeDef,
-			}, nil
-		} else {
-			value := unflattenConstantValue(typeDef, reltype.ConstantType.Value.Arguments)
-			cv, err := convertValue(typeDef, value)
-
-			typeDef["value"] = cv
-			return map[string]interface{}{
-				"type":  "Constant",
-				"value": typeDef,
-			}, err
-		}
-	}
-
-	if reltype.Tag == pb.Kind_PRIMITIVE_TYPE {
-		switch reltype.PrimitiveType {
-		case pb.PrimitiveType_STRING:
-			return _unmarshall(`{"type": "String"}`)
-		case pb.PrimitiveType_SYMBOL:
-			return _unmarshall(`{"type": "String"}`)
-		case pb.PrimitiveType_CHAR:
-			return _unmarshall(`{"type":"Char"}`)
-		case pb.PrimitiveType_BOOL:
-			return _unmarshall(`{"type":"Bool"}`)
-		case pb.PrimitiveType_INT_8:
-			return _unmarshall(`{"type":"Int8"}`)
-		case pb.PrimitiveType_INT_16:
-			return _unmarshall(`{"type":"Int16"}`)
-		case pb.PrimitiveType_INT_32:
-			return _unmarshall(`{"type":"Int32"}`)
-		case pb.PrimitiveType_INT_64:
-			return _unmarshall(`{"type":"Int64"}`)
-		case pb.PrimitiveType_INT_128:
-			return _unmarshall(`{"type":"Int128"}`)
-		case pb.PrimitiveType_UINT_8:
-			return _unmarshall(`{"type":"UInt8"}`)
-		case pb.PrimitiveType_UINT_16:
-			return _unmarshall(`{"type":"UInt16"}`)
-		case pb.PrimitiveType_UINT_32:
-			return _unmarshall(`{"type":"UInt32"}`)
-		case pb.PrimitiveType_UINT_64:
-			return _unmarshall(`{"type":"UInt64"}`)
-		case pb.PrimitiveType_UINT_128:
-			return _unmarshall(`{"type":"UInt128"}`)
-		case pb.PrimitiveType_FLOAT_16:
-			return _unmarshall(`{"type":"Float16"}`)
-		case pb.PrimitiveType_FLOAT_32:
-			return _unmarshall(`{"type":"Float32"}`)
-		case pb.PrimitiveType_FLOAT_64:
-			return _unmarshall(`{"type":"Float64"}`)
-		default:
-			panic(fmt.Sprintf("unhandled rel primitive type %v", reltype.PrimitiveType))
-		}
-	}
-
-	// check if reltype.ValueType is not empty
-	if reltype.Tag == pb.Kind_VALUE_TYPE {
-		var typeDefs []interface{}
-		for _, t := range reltype.ValueType.ArgumentTypes {
-			m, err := getColDefFromProtobuf(t)
-			if err != nil {
-				panic(err)
-			}
-
-			typeDefs = append(typeDefs, m)
-		}
-
-		typeDef := map[string]interface{}{
-			"type":     "ValueType",
-			"typeDefs": typeDefs,
-		}
-
-		return mapValueType(typeDef)
-	}
-
-	return _unmarshall(`{"type":"Unknown"}`)
-}
-
-func getColDefsFromProtobuf(relation pb.RelationId) []ColumnDef {
-	colDefs := make([]ColumnDef, 0)
-
-	arrowIndex := 0
-	for _, relType := range relation.Arguments {
-		typeDef, err := getColDefFromProtobuf(relType)
-		if err != nil {
-			panic(err)
-		}
-
-		colDef := new(ColumnDef)
-		colDef.TypeDef = typeDef
-		colDef.Metadata = *relType
-		if typeDef["type"] != "Constant" {
-			colDef.ArrowIndex = arrowIndex
-			arrowIndex++
-		}
-
-		colDefs = append(colDefs, *colDef)
-	}
-
-	return colDefs
-}
 
 func NewResultTable(relation rai.ArrowRelation) *ResultTable {
 	rs := new(ResultTable)
@@ -397,15 +93,6 @@ func (r *ResultTable) ToArrayRow() ([][]interface{}, error) {
 	}
 
 	keys := sortedMapKeys(m)
-
-	// for i := 0; i < int(r.RowsCount()); i++ {
-	// 	var row []interface{}
-	// 	for _, k := range keys {
-	// 		row = append(row, m[k][i])
-	// 	}
-
-	// 	out = append(out, row)
-	// }
 
 	for i := 0; i < int(r.RowsCount()); i++ {
 		var row []interface{}
@@ -538,14 +225,6 @@ func (r *ResultTable) Values() ([][]interface{}, error) {
 
 		out = append(out, v)
 	}
-
-	// for i := 0; i < int(r.RowsCount()); i++ {
-	// 	v, err := r.Get(i)
-	// 	if err != nil {
-	// 		return nil, err
-	// 	}
-	// 	out = append(out, v)
-	// }
 
 	return out, nil
 }
